Refactor: split lib into two modules.
